+++ server/src/middlewares/HandleInvestigationRequest.ts
